Backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.11
bcrypt==5.0.0
python-dotenv==1.2.1
plaid-python==38.3.0
cryptography==46.0.5
gunicorn==23.0.0
